requirements.txt
aiomysql==0.2.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-sqlalchemy==1.0.0
asyncpg==0.30.0
black==24.10.0
certifi==2024.12.14
click==8.1.8
dishka==1.4.2
dnspython==2.7.0
email_validator==2.2.0
-e git+ssh://git@github.com/pavelepanov/exam_tutor.git@85e3477c4de097cee8880980d7fc21aa2f4aeb2b#egg=exam_tutor
fastapi==0.115.6
fastapi-cli==0.0.7
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==5.13.2
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
netifaces==0.11.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
-e git+ssh://git@github.com/pavelepanov/solid_auth.git@c13b402fc464f28fa17bae16c5133d5fc2d7ac71#egg=solid_auth
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
